Adds a very simplistic one-way sql migration system.
Exposes two new symbols: *migration-table-name* and migrations.

The migrations functions accepts a list of strings that will modify
the db, in the order they are provided.  The *migration-table-name*
stores meta-data about the migrations, notable the md5 hash of the migration,
the migration query, and when it was run.

Future invocations check the *migration-table-name* for a matching hash and
only run the query if that's not found.

migrations is meant to be run at application start up or deployment time.

refs gainesville-green:#809 (.25)
refs ADWolf:#675 (.25)